refactor: change file name - Chat template 파일 이름 변경
